Added fetching user from token

+++ part4/blog-list/controllers/blogs.js
@@ -1,6 +1,7 @@
 const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1, id: 1 });
@@ -15,7 +16,9 @@ blogsRouter.post("/", async (request, response) => {
     return response.status(401).json({ error: "token missing or invalid" });
   }
 
-  const blog = new Blog(request.body);
+  const user = await User.findById(decodedToken.id);
+
+  const blog = new Blog({ ...request.body, author: user.name });
 
   const savedBlog = await blog.save();
   response.status(201).json(savedBlog.toJSON());
@@ -33,6 +36,13 @@ blogsRouter.put("/:blogId", async ({ body, params }, response) => {
 });
 
 blogsRouter.delete("/:blogId", async (request, response) => {
+  const { token } = request;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
   await Blog.findByIdAndRemove(request.params.blogId);
 
   response.status(204).end();
